refactored removed duplication: rules & config
