+++ ui/src/main/webapp/src/app/analysis-context/analysis-context.module.ts
@@ -18,12 +18,11 @@ import {ActionModule} from "patternfly-ng/action";
 import {TabsModule, TabsetConfig} from "ngx-bootstrap/tabs";
 import {ModalModule} from "ngx-bootstrap/modal";
 import {CustomRulesComponent} from "./custom-rules/custom-rules.component";
-import {UploadedRulesPathComponent} from "./custom-rules/uploaded-rules-path.component";
-import {ServerPathRulesComponent} from "./custom-rules/server-path-rules.component";
-import {UploadedRulePathModalComponent} from "./custom-rules/uploaded-rule-path-modal.component";
-import {ServerPathRulesModalComponent} from "./custom-rules/server-path-rules-modal.component";
-import {RulesListComponent} from "./custom-rules/rules-list.component";
-import {RulesComponent} from "./custom-rules/rules.component";
+import {UploadedRulesPathComponent} from "./custom-rules/upload/uploaded-rules-path.component";
+import {ServerPathRulesComponent} from "./custom-rules/server-path/server-path-rules.component";
+import {UploadedRulePathModalComponent} from "./custom-rules/upload/modal/uploaded-rule-path-modal.component";
+import {ServerPathRulesModalComponent} from "./custom-rules/server-path/modal/server-path-rules-modal.component";
+import {ConfigurationModule} from "../configuration/configuration.module";
 
 @NgModule({
     imports: [
@@ -34,7 +33,8 @@ import {RulesComponent} from "./custom-rules/rules.component";
         ToolbarModule,
         ActionModule,
         TabsModule,
-        ModalModule
+        ModalModule,
+        ConfigurationModule
     ],
     declarations: [
         AnalysisContextFormComponent,
@@ -50,9 +50,7 @@ import {RulesComponent} from "./custom-rules/rules.component";
         UploadedRulesPathComponent,
         ServerPathRulesComponent,
         UploadedRulePathModalComponent,
-        ServerPathRulesModalComponent,
-        RulesListComponent,
-        RulesComponent
+        ServerPathRulesModalComponent
     ],
     exports: [
         AnalysisContextFormComponent,
@@ -62,10 +60,7 @@ import {RulesComponent} from "./custom-rules/rules.component";
 
         SelectPackagesComponent,
         SelectPackagesWrapperComponent,
-        SelectPackagesSummaryComponent,
-
-        RulesListComponent,
-        RulesComponent
+        SelectPackagesSummaryComponent
     ],
     providers: [
         AnalysisContextService,
