--- ui/src/main/webapp/src/app/analysis-context/custom-rules/rules.component.html
@@ -1,28 +1,0 @@
-<ng-container *ngIf="ruleProvider.rules.length > 0; else elseNoRules">
-    <ng-container *ngIf="ruleProvider.rules.length > 1">
-        <label>Go to rule:</label>
-        <select #selectedRule
-                class="form-control"
-                name="ruleSelect"
-                style="width: auto; display: inline;"
-                (change)="scrollToRule(selectedRule.value)"
-                id="select-{{ruleProvider.id}}">
-            <option value="" disabled selected hidden>Select a rule</option>
-            <option 
-                *ngFor="let ruleEntity of ruleProvider.rules; let i = index"
-                [value]="ruleEntity.id">{{getLabelForRuleID(ruleEntity.ruleID, ruleProvider.providerID, i)}}</option>
-        </select>
-    </ng-container>
-    <div *ngFor="let ruleEntity of ruleProvider.rules; let i = index">
-        <h4 id="{{ruleEntity.id}}">
-            {{getLabelForRuleID(ruleEntity.ruleID, ruleProvider.providerID, i)}}
-            <i class="fa fa-angle-up fa-pull-right fa-border" aria-hidden="true"
-                (click)="scrollToRuleSetHeader(ruleProvider.id)" title="Go to rules set header"></i>
-        </h4>
-        <pre (load)="prettyPrint()" class="prettyprint"
-            style="white-space: pre-wrap">{{ruleEntity.ruleContents}}</pre>
-    </div>
-</ng-container>
-<ng-template #elseNoRules>
-    <span style="font-style: italic">No Rules Defined</span>
-</ng-template>
