--- ui/src/main/webapp/src/app/analysis-context/custom-rules/server-path-rules-modal.component.html
@@ -1,17 +1,0 @@
-<div class="modal-header">
-    <h4 class="modal-title pull-left">{{rulePath.path}}</h4>
-    <button type="button" class="close pull-right" aria-label="Close" (click)="bsModalRef.hide()">
-        <span aria-hidden="true">&times;</span>
-    </button>
-</div>
-<div #modalBody class="modal-body" style="height: 390px; overflow-y: auto;">
-    <wu-rules-list
-        [rulePath]="rulePath"
-        [ruleProviders]="ruleProviders"
-        [container]="modalBody"
-        [offset]="80">
-    </wu-rules-list>
-</div>
-<div class="modal-footer">
-    <button type="button" class="btn btn-default" (click)="bsModalRef.hide()">Close</button>
-</div>
