--- ui/src/main/webapp/src/app/analysis-context/custom-rules/server-path-rules-modal.component.scss
@@ -1,5 +1,0 @@
-@media (min-width: 992px) {
-    .list-view-pf-additional-info {
-        width: 12% !important;
-    }
-}
