+++ ui/src/main/webapp/src/app/analysis-context/custom-rules/server-path/modal/server-path-rules-modal.component.html
@@ -1,0 +1,40 @@
+<div class="modal-header">
+    <h4 class="modal-title pull-left">{{rulePath.path}}</h4>
+    <button type="button" class="close pull-right" aria-label="Close" (click)="bsModalRef.hide()">
+        <span aria-hidden="true">&times;</span>
+    </button>
+</div>
+<div #modalBody class="modal-body" style="height: 390px; overflow-y: auto;">
+    
+    <ng-template *ngIf="ruleProviders && ruleProviders.length > 0;then showList; else notShowList">
+    </ng-template>
+
+    <ng-template #showList>
+        <wu-rules-list
+            [rulePath]="rulePath"
+            [ruleProviders]="ruleProviders"
+            [container]="modalBody"
+            [offset]="80">
+        </wu-rules-list>
+    </ng-template>
+
+    <ng-template #notShowList>
+        <div class="blank-slate-pf" style="border: 0; background-color: transparent;">
+            <div class="blank-slate-pf-icon">
+                <span class="pficon pficon-info"></span>
+            </div>
+            <h1>
+                No Rules Paths found
+            </h1>
+            <p>
+                You can create your own custom rules and upload them here.
+            </p>
+            <p>
+                To learn more, visit the <a href="https://access.redhat.com/documentation/en-us/red_hat_application_migration_toolkit/" target="_blank">documentation</a>.
+            </p>
+        </div>
+    </ng-template>
+</div>
+<div class="modal-footer">
+    <button type="button" class="btn btn-default" (click)="bsModalRef.hide()">Close</button>
+</div>
