+++ ui/src/main/webapp/src/app/analysis-context/custom-rules/server-path/server-path-rules.component.scss
