+++ ui/src/main/webapp/src/app/analysis-context/custom-rules/upload/modal/uploaded-rule-path-modal.component.html
