+++ ui/src/main/webapp/src/app/analysis-context/custom-rules/upload/modal/uploaded-rule-path-modal.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { BsModalRef } from 'ngx-bootstrap/modal';
-import { RuleProviderEntity } from '../../generated/windup-services';
+import { RuleProviderEntity } from '../../../../generated/windup-services';
 
 @Component({
     selector: 'wu-uploaded-rule-path-modal',
