+++ ui/src/main/webapp/src/app/analysis-context/custom-rules/upload/uploaded-rules-path.component.html
