+++ ui/src/main/webapp/src/app/configuration/configuration.component.html
@@ -81,90 +81,11 @@
                     <a href="javascript:void(0)" (click)="removeFilters()" i18n>Remove</a> <ng-container i18n>the filter.</ng-container>
                 </div>
                 <ng-template #rulesList>
-                    <div class="rulesList list-group list-view-pf list-view-pf-view" *ngIf="hasFileBasedProviders(rulePath)" style="margin-top: 0px; border: none;">
-                        <div *ngFor="let ruleProvider of getFilteredRuleProvidersByPath(rulePath)" class="list-group-item" id="group-item-{{ruleProvider.id}}">
-                            <div class="list-group-item-header" (click)="clickHeader($event, ruleProvider)">
-                                <div class="list-view-pf-expand">
-                                    <span class="fa fa-angle-right" id="span-{{ruleProvider.id}}" [style.marginTop.px]="getRuleProviderMarginTop(ruleProvider)"></span>
-                                </div>
-                                <!-- If we need a checkbox for actions / global disabling.
-                                <div class="list-view-pf-checkbox">
-                                    <input type="checkbox">
-                                </div>
-                                -->
-                                <div class="list-view-pf-actions" style="margin-left: 5px;">
-                                    <!-- I just comment this button in case we'll use it again
-                                    <button class="btn btn-default" [disabled]="ruleProvider.rules.length < 1" (click)="displayRules(ruleProvider, $event)" i18n="Button">Show rules</button>
-                                    -->
-                                </div>
-                                <div class="list-view-pf-main-info">
-                                    <div class="list-view-pf-left">
-                                        <span class="fa fa-puzzle-piece list-view-pf-icon-sm"></span>
-                                    </div>
-                                    <div class="list-view-pf-body">
-                                        <div class="list-view-pf-description">
-                                            <div class="list-group-item-heading"> {{ruleProvider.providerID}} </div>
-                                            <div class="list-group-item-text">
-                                                <div style="width: 16pt; position: absolute;">
-                                                    <span class="fa fa-file-code-o list-view-pf-icon-md" style="font-size: 16pt;"></span>
-                                                </div>
-                                                <div style="margin-left: 25pt; ">
-                                                    <div>{{ruleProvider.origin}}</div>
-                                                    <div *ngIf="ruleProvider.loadError" style="font-weight: bold">
-                                                        {{ruleProvider.loadError}}
-                                                    </div>
-
-                                                    <div *ngIf="ruleProvider.sources.length > 0">
-                                                        Source technologies:
-                                                        <wu-technology *ngFor="let source of ruleProvider.sources" [technology]="source" class="techLabel"></wu-technology>
-                                                    </div>
-                                                    <div *ngIf="ruleProvider.targets.length > 0">
-                                                        Target technologies:
-                                                        <wu-technology *ngFor="let target of ruleProvider.targets" [technology]="target" class="techLabel"></wu-technology>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                        </div>
-                                        <div class="list-view-pf-additional-info">
-                                            <div class="list-view-pf-additional-info-item">
-                                                <span class="fa fa-list-ol"></span>
-                                                <strong>{{ruleProvider.rules ? ruleProvider.rules.length : 0}}</strong> <span i18n="Count, e.g. 5 Rules">Rules</span>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-                            <div class="list-group-item-container container-fluid hidden" id="container-{{ruleProvider.id}}" style="padding-top: 0px;">
-                                <div class="row">
-                                    <div class="modal-body">
-                                        <ng-container *ngIf="ruleProvider.rules.length > 0; else elseNoRules">
-                                            <ng-container *ngIf="ruleProvider.rules.length > 1">
-                                                <label>Go to rule:</label>
-                                                <select #selectedRule class="form-control" name="ruleSelect" style="width: auto; display: inline;"
-                                                        (change)="scrollToRule(selectedRule.value)"
-                                                        id="select-{{ruleProvider.id}}">
-                                                    <option value="" disabled selected hidden>Select a rule</option>
-                                                    <option *ngFor="let ruleEntity of ruleProvider.rules; let i = index" [value]="ruleEntity.id">{{getLabelForRuleID(ruleEntity.ruleID, ruleProvider.providerID, i)}}</option>
-                                                </select>
-                                            </ng-container>
-                                            <div *ngFor="let ruleEntity of ruleProvider.rules; let i = index">
-                                                <h4 id="{{ruleEntity.id}}">{{getLabelForRuleID(ruleEntity.ruleID, ruleProvider.providerID, i)}}
-                                                    <i class="fa fa-angle-up fa-pull-right fa-border" aria-hidden="true"
-                                                       (click)="scrollToRuleSetHeader(ruleProvider.id)"
-                                                       title="Go to rules set header"></i>
-                                                </h4>
-                                                <pre (load)="prettyPrint()" class="prettyprint" style="white-space: pre-wrap">{{ruleEntity.ruleContents}}</pre>
-                                            </div>
-                                        </ng-container>
-                                        <ng-template #elseNoRules>
-                                            <span style="font-style: italic">No Rules Defined</span>
-                                        </ng-template>
-                                    </div>
-                                </div>
-                            </div>
-
-                        </div>
-                    </div>
+                    <wu-rules-list *ngIf="hasFileBasedProviders(rulePath)"
+                        [rulePath]="rulePath"
+                        [ruleProviders]="getRuleProvidersByPath(rulePath)"
+                    >
+                    </wu-rules-list>
                 </ng-template>
             </div>
         </div>
