+++ ui/src/main/webapp/src/app/configuration/configuration.component.scss
@@ -1,43 +1,3 @@
-.rulesList .techLabel {
-  display: inline-block;
-  border: 2px ridge white /*1px solid #9ed0f1*/
-;
-  background-color: #e0f7ff;
-  padding: 0.1ex 0.4em;
-  margin: 0.1ex 0.4ex 0;
-}
-
-.list-group-item:hover {
-  background-color: rgb(222, 243, 255);
-}
-
-.rulesList .list-view-pf-body {
-  flex-basis: 89%;
-}
-
-@media (min-width: 992px) {
-  .rulesList .list-view-pf .list-group-item-heading {
-    font-size: initial;
-  }
-}
-
-/* Disable this weird rule from PF. */
-@media (min-width: 992px) {
-  .rulesList .list-view-pf-additional-info {
-    width: 12%;
-  }
-  .rulesList .list-view-pf-description {
-    width: 88%;
-  }
-}
-
-@media (min-width: 992px) {
-  .list-view-pf-additional-info {
-    -ms-flex: 0 0 auto;
-    flex: 0 0 auto;
-  }
-}
-
 .action-button {
   order: 2;
   margin-left: 40px;
